requirements.txt
absl-py==2.3.1
# Editable install with no version control (AdelaiDet==0.2.0)
-e /mnt/d/school/ship/ship_ocr/deepsolo-main/deepsolo-main/deepsolo
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==3.0.1
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.14.3
black==21.4b2
bleach==6.1.0
cachetools==5.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
click @ file:///croot/click_1698129812380/work
cloudpickle @ file:///croot/cloudpickle_1721657346512/work
contourpy @ file:///opt/conda/conda-bld/contourpy_1663827406301/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
Cython @ file:///croot/cython_1727769790504/work
cytoolz @ file:///croot/cytoolz_1701723583781/work
dask @ file:///croot/dask-core_1683065217061/work
decorator==5.2.1
defusedxml==0.7.1
detectron2==0.6+cu111
docopt==0.6.2
editdistance==0.8.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.16.1
fonttools @ file:///croot/fonttools_1713551344105/work
fsspec @ file:///croot/fsspec_1724855583730/work
future==1.0.0
fvcore==0.1.5.post20221221
google-auth==2.43.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hf-xet==1.2.0
huggingface-hub==0.36.0
hydra-core==1.3.2
idna==3.11
ImageHash==4.3.2
imageio @ file:///croot/imageio_1707247282708/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib_resources==6.4.5
iopath==0.1.9
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
llvmlite==0.39.1
locket @ file:///opt/conda/conda-bld/locket_1652903118915/work
Markdown==3.7
MarkupSafe==2.1.5
matplotlib @ file:///croot/matplotlib-suite_1670466153205/work
matplotlib-inline==0.1.7
mistune==3.1.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
mypy_extensions==1.1.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx @ file:///croot/networkx_1690561992265/work
numba==0.56.4
numpy==1.23.5
oauthlib==3.3.1
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging @ file:///croot/packaging_1720101850331/work
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
partd @ file:///croot/partd_1698702562572/work
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
polars==1.8.2
Polygon3==3.0.9.1
portalocker==3.0.0
prompt_toolkit==3.0.52
protobuf==4.21.2
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pydot==4.0.1
Pygments==2.19.2
pyparsing @ file:///croot/pyparsing_1725041624202/work
python-dateutil @ file:///croot/python-dateutil_1716495738603/work
pytz==2025.2
PyWavelets @ file:///croot/pywavelets_1670425177960/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==27.1.0
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.20.1
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.15.0
scipy==1.10.1
seaborn==0.13.2
shapely @ file:///croot/shapely_1680636282912/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.7
stack-data==0.6.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
timm==0.4.12
tinycss2==1.2.1
toml==0.10.2
toolz @ file:///croot/toolz_1667464077321/work
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado @ file:///croot/tornado_1718740109488/work
tqdm==4.67.1
traitlets==5.14.3
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3==2.2.3
wcwidth==0.2.14
webencodings==0.5.1
Werkzeug==3.0.6
yacs==0.1.8
yarg==0.1.9
zipp @ file:///croot/zipp_1729012354496/work
